refactor(context): rename packer service
