+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "eddie",
-    "version": "1.0.1",
+    "version": "1.0.2",
     "main": "dist/main.js",
     "bin": {
         "eddie": "dist/main.js"
