+++ src/core/engine/engine.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import path from "path";
 import type { CliRuntimeOptions } from "../../config/types";
 import { ConfigService } from "../../config/loader";
-import { ContextService } from "../context/packer";
+import { ContextService } from "../context/context.service";
 import { ProviderFactory } from "../providers";
 import { ToolRegistryFactory } from "../tools/registry";
 import { builtinTools } from "../tools/builtin";
