+++ test/unit/context.test.ts
@@ -3,7 +3,7 @@ import { describe, it, expect, beforeAll, afterAll } from "vitest";
 import { Test, type TestingModule } from "@nestjs/testing";
 import fs from "fs/promises";
 import path from "path";
-import { ContextService } from "../../src/core/context/packer";
+import { ContextService } from "../../src/core/context";
 import { LoggerService } from "../../src/io/logger";
 
 const tmpDir = path.join(process.cwd(), "test-temp");
